+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/location/CarLocationRenderer.kt
@@ -1,7 +1,6 @@
 package com.mapbox.androidauto.location
 
 import com.mapbox.androidauto.internal.logAndroidAuto
-import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.extension.androidauto.MapboxCarMapObserver
 import com.mapbox.maps.extension.androidauto.MapboxCarMapSurface
 import com.mapbox.maps.plugin.locationcomponent.location
@@ -10,7 +9,6 @@ import com.mapbox.maps.plugin.locationcomponent.location
  * Create a simple 3d location puck. This class is demonstrating how to
  * create a renderer. To Create a new location experience, try creating a new class.
  */
-@OptIn(MapboxExperimental::class)
 class CarLocationRenderer : MapboxCarMapObserver {
 
     override fun onAttached(mapboxCarMapSurface: MapboxCarMapSurface) {